refactor(lib/keyring/test): Factor out common test-utilities
