+++ lib/keyring/test/common.h
@@ -1,0 +1,23 @@
+/* Copyright 2023 Rufus Maurice Schäfing (wirklichniemand, nichtsundniemand)
+ *
+ * This file is part of `gnome-keyring-decryptor` and thus licensed under the terms
+ * of the GNU General Public License Version 3.
+ * A copy of this license can be found in the project's `LICENSE`-file.
+ */
+
+#ifndef COMMON_H
+#define COMMON_H
+
+#include <stdlib.h>
+#include <setjmp.h>
+#include <stdarg.h>
+#include <cmocka.h>
+
+struct no_input_state {
+	int empty_fd;
+};
+
+int setup_no_input(void **state);
+int teardown_no_input(void **state);
+
+#endif
